all: Move all the internal code to 'int'
Before there was a combination between 'int' and 'float64' which was anoying to deal with.

Also ran 'go vet ./...' to fix few things and cleaned up some unused methods

+++ client/game/map.go
@@ -41,7 +41,7 @@ func (m *Map) Draw(screen *ebiten.Image) {
 	s := m.game.Camera.GetState().(CameraState)
 	op.GeoM.Scale(s.Zoom, s.Zoom)
 	inverseZoom := maxZoom - s.Zoom + zoomScale
-	screen.DrawImage(imagesCache.Get(m.game.Store.Map.GetImageKey()).SubImage(image.Rect(int(s.X), int(s.Y), int((s.X+s.W)*inverseZoom), int((s.Y+s.H)*inverseZoom))).(*ebiten.Image), op)
+	screen.DrawImage(imagesCache.Get(m.game.Store.Map.GetImageKey()).SubImage(image.Rect(s.X, s.Y, int(float64((s.X+s.W))*inverseZoom), int(float64((s.Y+s.H))*inverseZoom))).(*ebiten.Image), op)
 
 	cs := m.game.Camera.GetState().(CameraState)
 	cp := m.game.Store.Players.FindCurrent()
@@ -50,23 +50,23 @@ func (m *Map) Draw(screen *ebiten.Image) {
 	for i := x - 4; i <= x+(18*16)+3; i++ {
 		// We draw 3 lines so it's kind of **bold**
 		// and it's easier to see
-		screen.Set(int(i-cs.X), int(y-cs.Y-4), cutils.Green)
-		screen.Set(int(i-cs.X), int(y-cs.Y-3), cutils.Green)
-		screen.Set(int(i-cs.X), int(y-cs.Y-2), cutils.Green)
+		screen.Set(i-cs.X, y-cs.Y-4, cutils.Green)
+		screen.Set(i-cs.X, y-cs.Y-3, cutils.Green)
+		screen.Set(i-cs.X, y-cs.Y-2, cutils.Green)
 
-		screen.Set(int(i-cs.X), int((y+86*16)-cs.Y+3), cutils.Green)
-		screen.Set(int(i-cs.X), int((y+86*16)-cs.Y+2), cutils.Green)
-		screen.Set(int(i-cs.X), int((y+86*16)-cs.Y+1), cutils.Green)
+		screen.Set(i-cs.X, (y+86*16)-cs.Y+3, cutils.Green)
+		screen.Set(i-cs.X, (y+86*16)-cs.Y+2, cutils.Green)
+		screen.Set(i-cs.X, (y+86*16)-cs.Y+1, cutils.Green)
 	}
 
 	// Color Left and Right
 	for i := y - 1; i <= y+(86*16); i++ {
-		screen.Set(int(x-cs.X-4), int(i-cs.Y), cutils.Green)
-		screen.Set(int(x-cs.X-3), int(i-cs.Y), cutils.Green)
-		screen.Set(int(x-cs.X-2), int(i-cs.Y), cutils.Green)
+		screen.Set(x-cs.X-4, i-cs.Y, cutils.Green)
+		screen.Set(x-cs.X-3, i-cs.Y, cutils.Green)
+		screen.Set(x-cs.X-2, i-cs.Y, cutils.Green)
 
-		screen.Set(int((x+18*16)-cs.X+3), int(i-cs.Y), cutils.Green)
-		screen.Set(int((x+18*16)-cs.X+2), int(i-cs.Y), cutils.Green)
-		screen.Set(int((x+18*16)-cs.X+1), int(i-cs.Y), cutils.Green)
+		screen.Set((x+18*16)-cs.X+3, i-cs.Y, cutils.Green)
+		screen.Set((x+18*16)-cs.X+2, i-cs.Y, cutils.Green)
+		screen.Set((x+18*16)-cs.X+1, i-cs.Y, cutils.Green)
 	}
 }
